we defined the default tranmission order is that LSB should comes first. Say, a hexadeximal 0x01020304 should be transmitted in '04->03->02->01' this manner in the network transmisson.
for a little endian machine, a hexadeximal 01020304 is stored as

|addr + 3| 01
|addr + 2| 02
|addr + 1| 03
|addr    | 04
that is, the LSB is stored in the smaller address.
so when transmitting, just use the pointer directly, say
put(addr);
put(addr + 1);
put(addr + 2);
put(addr + 3);

for a big endian machine, a hexadeximal 01020304 is stored as
|addr + 3| 04
|addr + 2| 03
|addr + 1| 02
|addr    | 01
that is, the LSB is stored in the biggest address.
so when transmitting, we need to reverser the oreder, say
put(addr + 3);
put(addr + 2);
put(addr + 1);
put(addr);